Former-commit-id: 3c4ae9bd0aa92b94d87f6ed47259419302274b74

+++ backend/src/main/java/com/ssafy/api/model/MyTrophy.java
@@ -23,8 +23,8 @@ import lombok.ToString;
 public class MyTrophy {
 	@Id
 	@Column(name = "USER_ID")
-	private int userID;
+	private long userID;
 
 	@Column(name = "TROPHY_ID")
-	private int trophyID;
+	private long trophyID;
 }
